+++ src/i18n/en.json
@@ -1,4 +1,3 @@
 {
-    "home":"home",
-    "tripListingHeading":"tripListingHeading"
+    "home":"home"
 }
