+++ src/i18n/es.json
@@ -1,4 +1,3 @@
 {
-    "home":"hogar",
-    "tripListingHeading":"trup"
+    "home":"hogar"
 }
