+++ src/pages/TripsListingPage/TripsListingPage.tsx
@@ -1,13 +1,11 @@
-import { useTranslation } from "react-i18next";
 import { TripsListingPageWrapper } from "./TripsListingPage.styled";
 import { TripCard } from "./components";
 
 export const TripsListingPage = () => {
-  const { t } = useTranslation();
 
   return (
     <TripsListingPageWrapper>
-      <h1>{t("tripListingHeading")}</h1>
+      <h1>Trip Listing Page</h1>
       <TripCard />
       <TripCard />
     </TripsListingPageWrapper>
